fix(stats): forecast decode crash + compaction over-thin on manual rerun
Review rework on PR #79 (issue #47).

1. `storage_forecast`'s total_bytes query 500'd on every single call:
   `sum(bigint)` returns Postgres `numeric`, not `bigint`, and sqlx refused
   to decode it into `i64` ("mismatched types ... NUMERIC"). Added the
   missing `::bigint` cast. The other two numeric fields in the same
   function were already correct (explicit cast / plain `count(*)`).
   Verified live: was a hard 500 on every request, now returns sane
   numbers.

2. Manually re-triggering `stats_compaction` (the admin Jobs page's
   "Run now", `system.jobs.update` — shipped in #40, reused here)
   compounded the COMPACTION_TIERS downsampling past the intended
   single-pass ratio, because each invocation's boundary math depends
   only on `Utc::now()` with no guard against being called twice before
   real time advances past the hour-wide slice. Reproduced live: 240
   seeded rows at the tier-7 boundary -> 60 (correct, first trigger) -> 15
   -> 4 -> 1 across four more manual triggers seconds apart -- the
   >99% loss is irreversible (`delete`, not a soft mark).

   Added an in-memory hour-bucket guard (`AtomicI64`, same idiom already
   used elsewhere in jobs.rs for last-refreshed tracking): the tiered
   passes run at most once per `STATS_COMPACTION_INTERVAL`-wide bucket no
   matter how many times the job is invoked within it (scheduled tick or
   manual trigger); a genuinely new bucket still runs normally. Verified
   live: three manual triggers in a row now leave a freshly-seeded
   tier-7-boundary band untouched (240 survivors all three times) instead
   of compounding. Numerically verified the swap-based dedup logic
   against a bucket sequence (same-bucket repeats all no-op; a new
   bucket is due exactly once) since the backend's own startup tick
   claims the current bucket before a live repro can observe the
   first-call path.

   Caveat, disclosed rather than silently left: the guard is in-memory,
   not persisted, so it resets on restart -- a rapid restart-loop within
   the same hour could still reprocess that hour's bucket once per
   restart. Far narrower than the original bug (any double-click), but
   not airtight; a persisted last-processed-boundary would close it
   fully if that matters here.

Not addressed (flagged, not blocking): the issue's "mark it event vs.
manual" save distinction is fully supported backend-side
(`SaveCaptureRequest.event_id`, validated + stored + joined) but
`replay.tsx` never supplies `event_id` -- the page carries no event
context, so every capture saved through the shipped UI is unconditionally
manual. Product/scope call, not a bug I'm guessing at the fix for.

fmt / clippy / 152 tests (0 new -- this repo's repo layer has no
`#[sqlx::test]` infra, see #74) all green. Both fixes verified against a
live Postgres instance with seeded data, not just read.

+++ backend/src/repos/stats.rs
@@ -749,8 +749,11 @@ pub async fn downsample_positions(
 /// headline numbers only (it deliberately does not model the compaction ladder's ongoing
 /// thinning, so the projections are an upper bound, not a precise forecast).
 pub async fn storage_forecast(pool: &PgPool) -> Result<StorageForecastBody, ApiError> {
+    // `sum(bigint)` returns `numeric` in Postgres (headroom against overflow), not `bigint` —
+    // cast back or sqlx refuses to decode it into `i64` (this 500'd on every call until caught in
+    // review: decode error "Rust type `i64` ... is not compatible with SQL type `NUMERIC`").
     let total_bytes: i64 = sqlx::query_scalar(
-        "select coalesce(sum(pg_total_relation_size(format('stats.%I', tablename)::regclass)), 0)
+        "select coalesce(sum(pg_total_relation_size(format('stats.%I', tablename)::regclass)), 0)::bigint
          from pg_tables where schemaname = 'stats'",
     )
     .fetch_one(pool)
